v.0.3

+++ action.py
@@ -27,6 +27,11 @@ from calibre_plugins.kindleunpack_plugin.dialogs import ProgressDialog, ResultsS
 from calibre_plugins.kindleunpack_plugin.utilities import (get_icon, KindleFormats, set_plugin_icon_resources,
                                 showErrorDlg, create_menu_item, create_menu_action_unique, build_log)
 
+# pulls in translation files for _() strings
+try:
+    load_translations()
+except NameError:
+    pass # load_translations() added in calibre 1.9
 
 class InterfacePlugin(InterfaceAction):
     name = 'KindleUnpack'
@@ -61,20 +66,20 @@ class InterfacePlugin(InterfaceAction):
         m = self.menu
         m.clear()
 
-        tool_tip = 'Convert the KF8 portions of the selected AZW3s to ePubs and add them to their respective books.'
+        tool_tip = _('Convert the KF8 portions of the selected AZW3s to ePubs and add them to their respective books.')
         create_menu_action_unique(self, m, _('KF8 to ePubs')+'...', 'mimetypes/epub.png', _(tool_tip),
                                                  False, triggered=partial(self.multi_dispatcher, book_ids, u'AZW3'))
 
-        tool_tip = 'Convert the KF8 portions of the selected AZW3s to ZIPs and add them to their respective books.'
+        tool_tip = _('Convert the KF8 portions of the selected AZW3s to ZIPs and add them to their respective books.')
         create_menu_action_unique(self, m, _('KF8 to ZIPs')+'...', 'mimetypes/zip.png', _(tool_tip),
                                                  False, triggered=partial(self.multi_dispatcher, book_ids, u'ZIP'))
 
-        tool_tip = 'Extract the PDFs from the AZW4 formats and add them to their respective books.'
+        tool_tip = _('Extract the PDFs from the AZW4 formats and add them to their respective books.')
         create_menu_action_unique(self, m, _('Extract PDFs')+'...', 'mimetypes/pdf.png', _(tool_tip),
                                                  False, triggered=partial(self.multi_dispatcher, book_ids, u'AZW4'))
 
         m.addSeparator()
-        tool_tip = 'Configure the KindleUnpack plugin\'s settings.'
+        tool_tip = _('Configure the KindleUnpack plugin\'s settings.')
         create_menu_action_unique(self, m, _('Customize plugin')+'...', 'config.png', _(tool_tip),
                                   None, triggered=self.show_configuration)
         self.gui.keyboard.finalize()
@@ -89,11 +94,11 @@ class InterfacePlugin(InterfaceAction):
         kindle_formats = ['MOBI', 'AZW', 'AZW3', 'AZW4', 'PRC']
         book_list = self.gatherKindleFormats([book_id], kindle_formats)
         if not book_list:
-            tool_tip = 'No suitable format to unpack.'
+            tool_tip = _('No suitable format to unpack.')
             error_menu = create_menu_item(self, m, _(tool_tip)+'...', None, _(tool_tip), None, None)
             error_menu.setEnabled(False)
             m.addSeparator()
-            tool_tip = 'Configure the KindleUnpack plugin\'s settings.'
+            tool_tip = _('Configure the KindleUnpack plugin\'s settings.')
             create_menu_action_unique(self, m, _('Customize plugin')+'...', 'config.png', _(tool_tip),
                                       None, triggered=self.show_configuration)
             self.gui.keyboard.finalize()
@@ -104,24 +109,24 @@ class InterfacePlugin(InterfaceAction):
             format_details = format_dict[format].get_format_details()
             # Weird and unlikely possiblity that there is no file on disk for this format at this point.
             if format_details['errors'] is not None and format_details['errors'] == 'path':
-                tool_tip = 'No file on disk. Can\'t unpack.'
-                mnu_msg = '{0} Has no file associated with it.'.format(format)
+                tool_tip = _('No file on disk. Can\'t unpack.')
+                mnu_msg = _('{0} Has no file associated with it.').format(format)
                 error_menu = create_menu_item(self, m, _(mnu_msg)+'...', None, _(tool_tip),
                                             None, None)
                 error_menu.setEnabled(False)
                 continue
             # Topaz format. Punt.
             elif format_details['errors'] is not None and format_details['errors'] == 'topaz':
-                tool_tip = 'Can\'t unpack Topaz books.'
-                mnu_msg = '{0} format is a Topaz book. Can\'t unpack.'.format(format)
+                tool_tip = _('Can\'t unpack Topaz books.')
+                mnu_msg = _('{0} format is a Topaz book. Can\'t unpack.').format(format)
                 error_menu = create_menu_item(self, m, _(mnu_msg)+'...', None, _(tool_tip),
                                         None, None)
                 error_menu.setEnabled(False)
                 continue
             # Unknown error. Very likely not a valid kindlebook file. Exact error found in format_details['errors'].
             elif format_details['errors'] is not None:
-                tool_tip = 'Unknown issues with this format.'
-                mnu_msg = '{0} format might not be a valid mobi/kindlebook.'.format(format)
+                tool_tip = _('Unknown issues with this format.')
+                mnu_msg = _('{0} format might not be a valid mobi/kindlebook.').format(format)
                 error_menu = create_menu_item(self, m, _(mnu_msg)+'...', None, _(tool_tip),
                                             None, None)
                 error_menu.setEnabled(False)
@@ -130,37 +135,37 @@ class InterfacePlugin(InterfaceAction):
             kindle_obj = format_details['kindle_obj']
 
             mnu_img = 'drm-unlocked.png'
-            mnu_tip = 'This {0} file is DRM-Free.'.format(format)
+            mnu_tip = _('This {0} file is DRM-Free.').format(format)
             if kindle_obj.isEncrypted:
                 print('isEncrypted = {0}'.format(kindle_obj.isEncrypted))
-                mnu_tip = 'This {0} file has DRM... can\'t unpack.'.format(format)
+                mnu_tip = _('This {0} file has DRM... can\'t unpack.').format(format)
                 mnu_img = 'drm-locked.png'
             ac = create_menu_item(self, m, _(format), mnu_img, _(mnu_tip), None)
             sm = QMenu()
             ac.setMenu(sm)
             # Standard unpack to external folder ... disable menu if kindlebook encrypted.
-            tool_tip = 'Unpack the {0}\'s source components'.format(format)
-            unpack_menu = create_menu_action_unique(self, sm, _('Unpack')+' {0}'.format(format), 'images/explode3.png',
+            tool_tip = _('Unpack the {0}\'s source components').format(format)
+            unpack_menu = create_menu_action_unique(self, sm, _('Unpack {0}').format(format), 'images/explode3.png',
                                                 _(tool_tip), False, triggered=partial(self.unpack_ebook, kindle_obj))
             if kindle_obj.isEncrypted:
                 unpack_menu.setEnabled(False)
 
             # Extract PDF file from AZW4
             if kindle_obj.isPrintReplica:
-                tool_tip = 'Extract the PDF from the Print Replica format and add it to the library.'
+                tool_tip = _('Extract the PDF from the Print Replica format and add it to the library.')
                 create_menu_action_unique(self, sm, _('Extract PDF')+'...', 'mimetypes/pdf.png', _(tool_tip), False,
                                             triggered=partial(self.extract_element, kindle_obj, book_id, u'AZW4', False))
 
             # Offer to split kindlegen dual format output.
             if kindle_obj.isComboFile:
-                tool_tip = 'Split the combo KF8/MOBI file into its two components.'
+                tool_tip = _('Split the combo KF8/MOBI file into its two components.')
                 create_menu_action_unique(self, sm, _('Split KF8/MOBI')+'...', 'edit-cut.png', _(tool_tip),
                                             False, triggered=partial(self.combo_split, kindle_obj))
 
             # Extract ePub from the unpacked contents and add to current book's formats.
             convert_menu = None
             if kindle_obj.isKF8 or kindle_obj.isComboFile:
-                tool_tip = 'Convert standalone KF8 file to its original ePub.'
+                tool_tip = _('Convert standalone KF8 file to its original ePub.')
                 convert_menu = create_menu_action_unique(self, sm, _('KF8 to ePub')+'...', 'mimetypes/epub.png', _(tool_tip),
                                             False, triggered=partial(self.extract_element, kindle_obj, book_id, u'AZW3', False))
             if kindle_obj.isEncrypted and convert_menu is not None:
@@ -169,7 +174,7 @@ class InterfacePlugin(InterfaceAction):
             # Extract ZIP from the unpacked contents and add to current book's formats.
             convert_menu = None
             if kindle_obj.isKF8 or kindle_obj.isComboFile:
-                tool_tip = 'Convert standalone KF8 file to ZIP.'
+                tool_tip = _('Convert standalone KF8 file to ZIP.')
                 convert_menu = create_menu_action_unique(self, sm, _('KF8 to ZIP')+'...', 'mimetypes/zip.png', _(tool_tip),
                                             False, triggered=partial(self.extract_element, kindle_obj, book_id, u'ZIP', False))
             if kindle_obj.isEncrypted and convert_menu is not None:
@@ -177,7 +182,7 @@ class InterfacePlugin(InterfaceAction):
 
         # Add menu item to go to plugin configuration.
         m.addSeparator()
-        tool_tip = 'Configure the KindleUnpack plugin\'s settings.'
+        tool_tip = _('Configure the KindleUnpack plugin\'s settings.')
         create_menu_action_unique(self, m, _('Customize plugin')+'...', 'config.png', _(tool_tip),
                                   None, triggered=self.show_configuration)
         self.gui.keyboard.finalize()
@@ -232,19 +237,19 @@ class InterfacePlugin(InterfaceAction):
         if target_format == 'AZW3':
             attr = 'isKF8'
             goal_format = 'EPUB'
-            status_msg_type='KF8 books'
-            action_type='Unpacking ePubs from'
+            status_msg_type=_('KF8 books')
+            action_type=_('Unpacking ePubs from')
         elif target_format == 'ZIP':
             target_format = 'AZW3'
             attr = 'isKF8'
             goal_format = 'ZIP'
-            status_msg_type='KF8 books'
-            action_type='Unpacking ZIPs from'
+            status_msg_type=_('KF8 books')
+            action_type=_('Unpacking ZIPs from')
         elif target_format == 'AZW4':
             attr = 'isPrintReplica'
             goal_format = 'PDF'
-            status_msg_type='Print Replica books'
-            action_type='Extracting PDFs from'
+            status_msg_type=_('Print Replica books')
+            action_type=_('Extracting PDFs from')
         books_info = self.gatherKindleFormats(book_ids, [target_format], goal_format)
         # If we have stuff ... send it on its way to the pretty ProgressDialog.
         if books_info:
@@ -260,14 +265,14 @@ class InterfacePlugin(InterfaceAction):
                 self.highlight_entries(ids_to_highlight)
             title = PLUGIN_NAME + ' v' + PLUGIN_VERSION
             plural = '' if len(successes) == 1 else 's'
-            msg = '<p>{0} {2} format{3} added to library. {1} not added. See log for details'.format(len(successes), len(failures), goal_format, plural)
+            msg = _('<p>{0} {2} format{3} added to library. {1} not added. See log for details').format(len(successes), len(failures), goal_format, plural)
             log = build_log(failures, successes, target_format, goal_format, status_msg_type[:-1])
             # print (log)
             sd = ResultsSummaryDialog(self.gui, title, msg, log)
             sd.exec_()
         else:
             return info_dialog(None, _(PLUGIN_NAME + ' v' + PLUGIN_VERSION),
-                '<p>Nothing to do. Perhaps no books selected had {0} formats.'.format(target_format), show=True)
+                _('<p>Nothing to do. Perhaps no books selected had {0} formats.').format(target_format), show=True)
 
     def highlight_entries(self, ids_to_highlight):
         # self.gui.library_view.model().books_added(len(ids_to_highlight))
@@ -283,7 +288,7 @@ class InterfacePlugin(InterfaceAction):
         '''
         outdir = self.directoryChooser()
         if outdir:
-            kindle_obj.setContentDir(cfg.plugin_prefs['Kindle_Content_Folder'])
+            kindle_obj.setKindleContentDir(cfg.plugin_prefs['Kindle_Content_Folder'])
             try:
                 kindle_obj.unpackMOBI(outdir)
             except Exception as e:
@@ -296,7 +301,8 @@ class InterfacePlugin(InterfaceAction):
         '''
         outdir = PersistentTemporaryDirectory()
         errmsg = ''
-        kindle_obj.setContentDir(cfg.plugin_prefs['Kindle_Content_Folder'])
+        kindle_obj.setZipCompressType(cfg.plugin_prefs['Zip_Compress_Type'])
+        kindle_obj.setKindleContentDir(cfg.plugin_prefs['Kindle_Content_Folder'])
         if target == 'AZW3':
             errmsg = 'An'
             format = 'EPUB'
@@ -327,22 +333,23 @@ class InterfacePlugin(InterfaceAction):
 
         if os.path.exists(bookfile):
             if not self.update_db(bookfile, format, book_id):
-                errmsg += ' {0} format already exists for this book in this library! No attempt to overwrite it will be made.'.format(format)
+                errmsg += _(' {0} format already exists for this book in this library! No attempt to overwrite it will be made.').format(format)
                 if quiet:
                     return False, None
                 return showErrorDlg(errmsg, self.gui)
             current_idx = self.gui.library_view.currentIndex()
-            # remove outdir
-            shutil.rmtree(outdir)
+            # delete outdir
+            if cfg.plugin_prefs['Always_Delete_Temp_Files']:
+                shutil.rmtree(outdir)
             #
             if current_idx.isValid():
                 self.gui.library_view.model().current_changed(current_idx, current_idx)
             if quiet:
                 return True, None
             return info_dialog(None, _(PLUGIN_NAME + ' v' + PLUGIN_VERSION),
-            '<p>{0} successfully unpacked and added to ebook\'s formats in library.'.format(format), show=True)
+            _('<p>{0} successfully unpacked and added to ebook\'s formats in library.').format(format), show=True)
 
-        errmsg = 'Couldn\'t find {0} in unpacked kindlebook.'.format(format)
+        errmsg = _('Couldn\'t find {0} in unpacked kindlebook.').format(format)
         if quiet:
             return False, errmsg
         return showErrorDlg(errmsg, self.gui)
